Add CryptoPunks contract to ethAbi.json
Hi, I'm one of the creators of the CryptoPunks: http://www.larvalabs.com/cryptopunks We are getting a number of users who are new to Ethereum and I was hoping to reduce some steps by pointing them to a preprepared interface to the contract on MEW. I hope I've done this correctly, please let me know if not. Thanks!