Add SingularDTV derivation to mnemonic paths.
The HD wallet path for singular is differrent from all the rest, user is
required to always enter a custom path to get to his/hers SNGLS tokens
from a SingularDTV Lightwallet generated mnemonic.
The reason for this is that SingularDTV launched it's wallet before
`m/44'/60'/...` became standard.
Since a lot of users keep asking about the singular derivation path
when used in conjuction with MyEtherWallet, adding this path to the options
will help quite a few people.